Agents-Azure/requirements.txt
openai==1.97.1
python-dotenv==1.0.1
requests==2.32.4
azure-identity==1.19.0
pydantic==2.10.1
